alpha to teta

+++ MyDrivers/voiles.h
@@ -4,7 +4,7 @@
 #include "stm32f1xx_ll_gpio.h"
 
 int voiles_init();
-int voiles_alpha_to_teta(int alpha);
+float voiles_alpha_to_teta(int alpha);
 int voiles_teta_to_ms(int teta);
 
 #endif
